planning to readjust movement calculations for a kinematic rigidbody
So up until now, I knew I wanted a rigid body controller, however I was unaware of the significance when it came to the difference between Dynamic Rigid body and a Kinematic Rigidbody. I think the Kinematic approach will give the player more control over their movements while still being able to interact with other Dynamic Rigdbodys.

The most recent movement delay speed bug is what clued me onto this as I found that the Player was falling from the preset y=1.09 to y=1, making contact with the ground. Prior to the state manager, this had little effect on feeling since I was setting rb.velocity directly in as opposed to rb.addForce in MoveGround() in PlayerMovement.cs, which at the time would not work and I would not know why. However I found that when tapping the movement, wasd, keys in any direction for ~1 frame, the Update() would run however movement calculation in FixedUpdate() didnt have any affect. Why? Probably because the rb is touching the ground, and since we're accelerating it will take longer because of the extra friction. Keep in mind we are already using MoveGround() calculation to take into account friction. So this is just extra rb friction with our calculated friction.

I think levitating/floating/springing/hovering the character slightly above the ground will help remedy this, I just have to find a way to do it properly and adapt the current PlayerMovement calculations